Tweety Bot version 1
Tweety Bot version 1: Tweety Bot will scrape a website for the quote of
the day, this is written in QuoteParser.py. The website is
www.eduro.com. The web crawler is written in version 3.3. Tweety Bot
will write this quote to the txt file qotd.txt. Tweety Bot can then
take that quote using TweetyBot.py and post it to twitter, as long as
the quote is less than 140 chars. TweetyBot.py uses the tweepy module,
which must be downloaded. Tweepy needs Python 2.7, so TweetyBot.py is
written in 2.7. So you must run the two scripts separately.